fix(deploy): make the new tests reachable and split buildDeployment
Two Fast Quality Gates failures on #12400.

`check:test-discovery` reported all three new suites as orphans: the unit
runner's subdirectory glob is an explicit allowlist
(`tests/unit/{api,auth,authz,...}/**`) and `deploy` is not in it, so nothing
under `tests/unit/deploy/` was ever collected. The suites passed locally and in
no CI job — the exact failure mode that gate was added for after the 2026-06-09
audit found ~135 orphaned tests, one of them a Hard Rules #15/#17 check whose
asserts were already failing unnoticed.

Moved the three files to `tests/unit/` rather than widening the allowlist:
adding a directory there changes the runner for every contributor, which is not
this PR's business. Import depths adjusted; the fourth suite was already at the
collected level.

`check:complexity-ratchets` flagged `buildDeployment` at 83 lines against a
limit of 80. The pod template is almost entirely the container object, so that
object moved to `buildContainer`. No behaviour change — the generator's output
is byte-identical, which the parity test against deploy/kubernetes/base pins.

Both gates verified locally: test-discovery OK, complexityNewCode=0.
